Clean up React dead code and consolidate duplicate helpers
- Create lib/formatters.ts with shared formatDuration, formatDate,
  formatDateTime, formatNumber, and rankBadgeClass helpers
- Remove 3 duplicate formatDuration implementations across GameResults,
  GameSummary, and PlayerHistory
- Remove duplicate formatDate/rankBadge from Achievements and PlayerHistory
- Replace local vcBadge in Games.tsx with shared vcBadgeCss/vcText from
  lib/victory.ts
- Move formatNumber from PlayerResources to shared formatters
- Remove 16 unused TypeScript interfaces/types from api/types.ts
