+++ src/ReactClient/src/pages/GameSummary.tsx
@@ -4,16 +4,7 @@ import apiClient from '@/api/client'
 import type { GameResultsViewModel } from '@/api/types'
 import { PageLoader } from '@/components/PageLoader'
 import { ApiError } from '@/components/ApiError'
-
-function formatDuration(ms: number): string {
-  const totalSec = Math.floor(ms / 1000)
-  const days = Math.floor(totalSec / 86400)
-  const hours = Math.floor((totalSec % 86400) / 3600)
-  const mins = Math.floor((totalSec % 3600) / 60)
-  if (days >= 1) return `${days}d ${hours}h`
-  if (hours >= 1) return `${hours}h ${mins}m`
-  return `${mins}m`
-}
+import { formatDuration } from '@/lib/formatters'
 
 export function GameSummary() {
   const { gameId } = useParams<{ gameId: string }>()
