+++ src/ReactClient/src/pages/PlayerHistory.tsx
@@ -3,26 +3,7 @@ import apiClient from '@/api/client'
 import type { PlayerHistoryViewModel } from '@/api/types'
 import { PageLoader } from '@/components/PageLoader'
 import { ApiError } from '@/components/ApiError'
-
-function formatDuration(startIso: string, endIso: string): string {
-  const ms = new Date(endIso).getTime() - new Date(startIso).getTime()
-  if (ms <= 0) return '—'
-  const totalMinutes = Math.floor(ms / 60000)
-  const days = Math.floor(totalMinutes / 1440)
-  const hours = Math.floor((totalMinutes % 1440) / 60)
-  const minutes = totalMinutes % 60
-  if (days > 0) return `${days}d ${hours}h`
-  if (hours > 0) return `${hours}h ${minutes}m`
-  return `${minutes}m`
-}
-
-function rankBadgeClass(rank: number, players: number): string {
-  if (rank === 1) return 'bg-warning text-black'
-  if (rank === 2) return 'bg-muted-foreground text-black'
-  if (rank === 3) return 'bg-warning/60 text-white'
-  if (rank > players / 2) return 'bg-secondary text-secondary-foreground opacity-60'
-  return 'bg-secondary text-secondary-foreground'
-}
+import { formatDurationFromDates, rankBadgeClass } from '@/lib/formatters'
 
 export function PlayerHistory() {
   const { data, isLoading, error, refetch } = useQuery<PlayerHistoryViewModel>({
@@ -98,7 +79,7 @@ export function PlayerHistory() {
                   })}
                 </td>
                 <td className="px-3 py-2 text-muted-foreground hidden sm:table-cell">
-                  {formatDuration(entry.startTime, entry.endTime)}
+                  {formatDurationFromDates(entry.startTime, entry.endTime)}
                 </td>
                 <td className="px-3 py-2">
                   <span
